factor out the CRModifier
